% This was not Generated by roxygen2 (4.1.1): edit by hand. Changes in tauchen.R not reflected here. 
\name{Rtauchen}
\alias{Rtauchen}
\title{Rtauchen}
\usage{
gmapsdistance(ne, ssigma_eps, llambda_eps, m)
}
\arguments{
\item{ne}{Number of points in grid}

\item{ssigma_eps}{Standard deviation of exogenous shock}

\item{llambda_eps}{persistence parameter}

\item{m}{Tauchen parameter}
}
\value{
a matrix with the corresponding transition matrix
}
\description{
Discretize AR1 process
}
\details{
Discretize AR process
}
\examples{
results = TMatrix(2, 1.0e-5, 0.1,0.4)
results
}

