% This was not Generated by roxygen2 (4.1.1): edit by hand. Changes in tauchen.R not reflected here. 
\name{Rtauchen}
\alias{Rtauchen}
\title{Rtauchen}
\usage{
Rtauchen(ne, ssigma_eps, llambda_eps, m)
}
\arguments{
\item{ne}{Number of points of the grid of the finite-state Markov chain that mimics the AR(1) process}

\item{ssigma_eps}{Standard deviation of exogenous shock in the AR(1) process}

\item{llambda_eps}{Persistence parameter of the AR(1) process}

\item{m}{Tauchen parameter for the width of the process (number of standard deviations of the AR(1) process covered by the grid)}
}
\value{
A matrix with the corresponding to the transition matrix of the finite-state Markov chain that approximates the AR(1) process
}
\description{
This function generates a matrix of transition probabilites of a finite-state Markov chain that mimics an AR(1) process with persistence parameter llamda, standard deviation ssigma and a fixed parameter m.
}
\details{
See Tauchen (1986) for details.
}
\examples{
results = Rtauchen(2, 1.0e-5, 0.1,0.4)
results
}




